style: Move presentational hint synthesis into `legacy.rs`.
